% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_text.R
\name{classify_text}
\alias{classify_text}
\title{Classify Text With GPT-3}
\usage{
classify_text(
  text,
  instructions =
    "Classify the sentiment of the following text as Positive, Negative, or Neutral.",
  examples = NULL,
  model = "gpt-3.5-turbo",
  temperature = as.integer(0),
  max_tokens = as.integer(5),
  openai_api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{text}{The text to be classified}

\item{instructions}{Instructions to be included in the GPT prompt (format them like you would format instructions to a human research assistant).}

\item{examples}{A dataframe of "few-shot" examples. Must include one column called 'text' with the example text(s) and another column called "label" with the correct label(s).}

\item{model}{Which model variant of GPT to use. Defaults to 'gpt-3.5-turbo'}

\item{openai_api_key}{Your API key. By default, looks for a system environment variable called "OPENAI_API_KEY" (recommended option). Otherwise, it will prompt you to enter the API key as an argument.}
}
\value{
A classification of the text
}
\description{
Constructs a GPT-3 prompt to classify a string of text according to user instructions
}
\examples{
classify_text(text = 'I am so very sad today',
              instructions =
              'Classify the sentiment of the following string of text as either Positive, Negative, or Neutral.')
}
