% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_prompt.R
\name{complete_prompt}
\alias{complete_prompt}
\title{Complete a Custom GPT Prompt}
\usage{
complete_prompt(
  prompt,
  model = "gpt-3.5-turbo",
  openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  max_tokens = as.integer(1),
  temperature = as.integer(0)
)
}
\arguments{
\item{prompt}{The prompt to use as input for GPT}

\item{model}{Which model variant of GPT to use. Defaults to 'gpt-3.5-turbo'}

\item{openai_api_key}{Your API key. By default, looks for a system environment variable called "OPENAI_API_KEY" (recommended option). Otherwise, it will prompt you to enter the API key as an argument.}

\item{max_tokens}{How many tokens (roughly 4 characters of text) should GPT return? Defaults to a single token (next word prediction).}

\item{temperature}{A number between 0 and 2. When set to zero, GPT will always return the most probable next token. When set higher, GPT will select the next word probabilistically.}
}
\value{
A string of text generated by GPT.
}
\description{
Complete a Custom GPT Prompt
}
\examples{
complete_prompt('I feel like a')
complete_prompt('Write a haiku about frogs.', max_tokens = 100)
}
