% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scotus_tweets}
\alias{scotus_tweets}
\title{Tweets About The Supreme Court of the United States}
\format{
A data frame with 945 rows and 5 columns:
\describe{
\item{tweet_id}{A unique ID}
\item{text}{The text of the tweet}
\item{expert1, expert2, expert3}{Hand-coded sentiment score (-1 = negative, 0 = neutral, 1 = positive)}
}
}
\usage{
data(scotus_tweets)
}
\description{
This dataset contains 945 tweets referencing the Supreme Court,
collected between XXXX and YYYY, along with expert-coded sentiment
scores.
}
\examples{
data(scotus_tweets)
}
\references{
Ornstein et al. (2022). "How To Train Your Stochastic Parrot"
}
\keyword{datasets}
