% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scotus_tweets_examples}
\alias{scotus_tweets_examples}
\title{Labelled Example Tweets About The Supreme Court of the United States}
\format{
A data frame with 5 rows and 2 columns:
\describe{
\item{text}{The text of the tweet}
\item{label}{The "true" label (Positive, Negative, or Neutral)}
}
}
\usage{
data(scotus_tweets_examples)
}
\description{
This dataset contains 5 example tweets referencing the Supreme Court
along with a sentiment label. These can be used as few-shot learning
examples for classifying tweets in the \code{scotus_tweets} dataset.
}
\references{
Ornstein et al. (2022). "How To Train Your Stochastic Parrot"
}
\keyword{datasets}
