% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_openai.R
\name{setup_openai}
\alias{setup_openai}
\title{Setup OpenAI Python Module}
\usage{
setup_openai(method = "auto", conda = "auto")
}
\arguments{
\item{method}{Method for installing python package (see reticulate::py_install)}

\item{conda}{Use conda? (see reticulate::py_install)}
}
\description{
A helper function to install the OpenAI Python module, if it is not already on your system.
}
