% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_text.R
\name{classify_text}
\alias{classify_text}
\title{Classify Text With GPT-3}
\usage{
classify_text(
  text,
  instructions =
    "Classify the sentiment of the following tweet as Positive, Negative, or Neutral.",
  model = "text-davinci-003"
)
}
\arguments{
\item{text}{The text to be classified}

\item{instructions}{Instructions to be included in the GPT-3 prompt (format them like you would format instructions to a human research assistant).}

\item{model}{Which model variant of GPT-3 to use. Defaults to 'text-davinci-003'}
}
\value{
A dataframe with 5 classifications and their probabilities as assigned by GPT-3
}
\description{
Constructs a GPT-3 prompt to classify a string of text according to user instructions
}
\examples{
classify_text(text = 'I am so very sad today',
              instructions = 'Classify the sentiment of the following string of
                              text as either Positive, Negative, or Neutral.')
}
