% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_prompt.R
\name{complete_prompt}
\alias{complete_prompt}
\title{Complete a Custom GPT-3 Prompt}
\usage{
complete_prompt(
  prompt,
  model = "text-davinci-003",
  openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  max_tokens = 1,
  temperature = 0
)
}
\arguments{
\item{prompt}{The prompt to use as input for GPT-3}

\item{model}{Which model variant of GPT-3 to use. Defaults to 'text-davinci-003'}

\item{openai_api_key}{Your API key. By default, looks for a system environment variable called "OPENAI_API_KEY" (recommended option). Otherwise, it will prompt you to enter the API key as an argument.}

\item{max_tokens}{How many tokens (roughly 4 characters of text) should GPT-3 return? Defaults to a single token (next word prediction).}

\item{temperature}{A number between 0 and 100. When set to zero, GPT-3 will always return the most probable next token. When set higher, GPT-3 will select the next word probabilistically.}
}
\value{
If max_tokens is 1, returns a dataframe with the 5 most likely next words and their probabilities as assigned by GPT-3. If max_tokens > 1, returns a single string of text generated by GPT-3.
}
\description{
Complete a Custom GPT-3 Prompt
}
\examples{
complete_prompt('I feel like a')
complete_prompt('Write a haiku about frogs.', max_tokens = 100)
}
