% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_text.R
\name{parse_text}
\alias{parse_text}
\title{Create Dataframe From Text}
\usage{
parse_text(
  text,
  instructions,
  var_descriptions,
  col_names = character(0),
  model = "text-davinci-003",
  openai_api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{text}{The text to be converted into a tidy dataframe}

\item{instructions}{Instructions to be included in the GPT-3 prompt (format them like you would format instructions to a human research assistant, in as much detail as possible).}

\item{var_descriptions}{A list of descriptions for each variable in the dataframe.}

\item{col_names}{Optional: A list of column names for the resulting dataframe. If not provided, var_descriptions will be used for column names.}

\item{model}{Which model variant of GPT-3 to use. Defaults to 'text-davinci-003'}

\item{openai_api_key}{Your API key. By default, looks for a system environment variable called "OPENAI_API_KEY" (recommended option). Otherwise, it will prompt you to enter the API key as an argument.}
}
\value{
A dataframe with a number of columns equal to \code{ncol(var_descriptions)}
}
\description{
Constructs a GPT-3 prompt to parse data from unstructured text according to the user's instructions.
}
\examples{
parse_text(text = "John went to the store. Rosie ran errands at the mall.",
           instructions = "Create a dataset from the following passage with the names of each character and the place they went.",
           var_descriptions = c("Name of Character", "Where They Went"),
           col_names = c('name', 'destination') )
}
