% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_text.R
\name{parse_text}
\alias{parse_text}
\title{Create Dataframe From Text}
\usage{
parse_text(
  text,
  instructions =
    "Create a dataset from the following package extracting relevant information",
  var_descriptions = character(0),
  col_names = character(0),
  model = "gpt-3.5-turbo",
  openai_api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{text}{The text to be converted into a tidy dataframe}

\item{instructions}{Instructions to be included in the GPT prompt (format them like you would format instructions to a human research assistant, in as much detail as possible).}

\item{var_descriptions}{Optional: A list of descriptions for each variable in the dataframe. (If not specified, will query GPT to create these)}

\item{col_names}{Optional: A list of column names for the resulting dataframe. Must be the same length as the list of variable descriptions. If not provided, var_descriptions will be used for column names.}

\item{model}{Which model variant of GPT to use. Defaults to 'gpt-3.5-turbo'}

\item{openai_api_key}{Your API key. By default, looks for a system environment variable called "OPENAI_API_KEY" (recommended option). Otherwise, it will prompt you to enter the API key as an argument.}
}
\value{
A dataframe with a number of columns equal to \code{ncol(var_descriptions)}
}
\description{
Constructs a GPT prompt to parse data from unstructured text according to the user's instructions.
}
\examples{
parse_text(text = "John went to the store. Rosie ran errands at the mall.",
           instructions = "Create a dataset from the following passage with the names of each character and the place they went.",
           var_descriptions = c("Name of Character", "Where They Went"),
           col_names = c('name', 'destination') )
}
